+++ create-pr.sh
@@ -9,9 +9,10 @@ export CREATE_PR_ROOT_DIR
 
 source "$CREATE_PR_ROOT_DIR/src/dev/debug.sh"
 source "$CREATE_PR_ROOT_DIR/src/validation.sh"
+source "$CREATE_PR_ROOT_DIR/src/pr_ticket.sh"
 source "$CREATE_PR_ROOT_DIR/src/pr_body.sh"
 source "$CREATE_PR_ROOT_DIR/src/pr_label.sh"
-source "$CREATE_PR_ROOT_DIR/src/pr_format.sh"
+source "$CREATE_PR_ROOT_DIR/src/pr_title.sh"
 source "$CREATE_PR_ROOT_DIR/src/env_configuration.sh"
 source "$CREATE_PR_ROOT_DIR/src/console_header.sh"
 source "$CREATE_PR_ROOT_DIR/src/main.sh"