+++ src/env_configuration.sh
@@ -18,8 +18,8 @@ ASSIGNEE=${ASSIGNEE:-"@me"}
 
 BRANCH_NAME=$(git rev-parse --abbrev-ref HEAD 2>/dev/null) || error_and_exit "Failed to get the current branch name."
 LABEL=${LABEL:-$(pr_label "$BRANCH_NAME" "${PR_LABEL_MAPPING:-}")}
-PR_TITLE=$(format_title "$BRANCH_NAME")
-PR_BODY=$(format_pr_body "$BRANCH_NAME" "$PR_TEMPLATE")
+PR_TITLE=$(pr_title "$BRANCH_NAME")
+PR_BODY=$(pr_body "$BRANCH_NAME" "$PR_TEMPLATE")
 
 export BASE_BRANCH
 export ASSIGNEE