+++ tests/unit/pr_body_test.sh
@@ -6,10 +6,10 @@ function set_up() {
   source "$CREATE_PR_ROOT_DIR/src/pr_body.sh"
 }
 
-function test_format_pr_body_link_without_comment() {
+function test_pr_body_link_without_comment() {
   export PR_TICKET_LINK_PREFIX=https://your-ticket-system.com/
 
-  local actual=$(format_pr_body "TICKET-123-my_branch-with-1-number"\
+  local actual=$(pr_body "TICKET-123-my_branch-with-1-number"\
     "$CREATE_PR_ROOT_DIR/.github/PULL_REQUEST_TEMPLATE.md")
 
   # validate that the link is not inside a HTML comment
@@ -20,20 +20,20 @@ function test_format_pr_body_link_without_comment() {
   assert_contains "<!-- " "$actual"
 }
 
-function test_format_pr_body_link_with_PR_TICKET_LINK_PREFIX() {
+function test_pr_body_link_with_PR_TICKET_LINK_PREFIX() {
   export PR_TICKET_LINK_PREFIX=https://your-ticket-system.com/
 
-  local actual=$(format_pr_body "TICKET-123-my_branch-with-1-number"\
+  local actual=$(pr_body "TICKET-123-my_branch-with-1-number"\
     "$CREATE_PR_ROOT_DIR/.github/PULL_REQUEST_TEMPLATE.md")
 
   assert_contains "https://your-ticket-system.com/TICKET-123" "$actual"
   assert_not_contains "{{ TICKET_LINK }}" "$actual"
 }
 
-function test_format_pr_body_link_without_PR_TICKET_LINK_PREFIX() {
+function test_pr_body_link_without_PR_TICKET_LINK_PREFIX() {
   export PR_TICKET_LINK_PREFIX=
 
-  local actual=$(format_pr_body "TICKET-123-my_branch"\
+  local actual=$(pr_body "TICKET-123-my_branch"\
     "$CREATE_PR_ROOT_DIR/.github/PULL_REQUEST_TEMPLATE.md")
 
   assert_not_contains "https://your-ticket-system.com/" "$actual"
@@ -41,29 +41,29 @@ function test_format_pr_body_link_without_PR_TICKET_LINK_PREFIX() {
   assert_contains "Nope" "$actual"
 }
 
-function test_format_pr_body_link_without_ticket_key_default_link_prefix_text() {
+function test_pr_body_link_without_ticket_key_default_link_prefix_text() {
   export PR_TICKET_LINK_PREFIX=https://your-ticket-system.com/
 
-  local actual=$(format_pr_body "123-my_branch"\
+  local actual=$(pr_body "123-my_branch"\
     "$CREATE_PR_ROOT_DIR/.github/PULL_REQUEST_TEMPLATE.md")
 
   assert_contains "https://your-ticket-system.com/123" "$actual"
 }
 
-function test_format_pr_body_link_custom_link_prefix_text() {
+function test_pr_body_link_custom_link_prefix_text() {
   export PR_TICKET_LINK_PREFIX=https://your-ticket-system.com/
   export PR_TICKET_PREFIX_TEXT="Fixes: "
 
-  local actual=$(format_pr_body "123-my_branch"\
+  local actual=$(pr_body "123-my_branch"\
     "$CREATE_PR_ROOT_DIR/.github/PULL_REQUEST_TEMPLATE.md")
 
   assert_contains "Fixes: https://your-ticket-system.com/123" "$actual"
 }
 
-function test_format_pr_body_link_without_ticket_number() {
+function test_pr_body_link_without_ticket_number() {
   export PR_TICKET_LINK_PREFIX=https://your-ticket-system.com/
 
-  local actual=$(format_pr_body "TICKET-my_branch"\
+  local actual=$(pr_body "TICKET-my_branch"\
     "$CREATE_PR_ROOT_DIR/.github/PULL_REQUEST_TEMPLATE.md")
 
   assert_not_contains "https://your-ticket-system.com/" "$actual"
@@ -71,36 +71,36 @@ function test_format_pr_body_link_without_ticket_number() {
   assert_contains "Nope" "$actual"
 }
 
-function test_format_pr_body_link_without_pr_template() {
+function test_pr_body_link_without_pr_template() {
   export PR_TICKET_LINK_PREFIX=https://your-ticket-system.com/
 
-  local actual=$(format_pr_body "TICKET-123-my_branch" "")
+  local actual=$(pr_body "TICKET-123-my_branch" "")
 
   assert_same "" "$actual"
 }
 
-function test_format_pr_body_link_with_branch_with_numbers_no_prefix() {
+function test_pr_body_link_with_branch_with_numbers_no_prefix() {
   export PR_TICKET_LINK_PREFIX=https://your-ticket-system.com/
 
-  local actual=$(format_pr_body "TICKET-123-my-4-th-branch"\
+  local actual=$(pr_body "TICKET-123-my-4-th-branch"\
     "$CREATE_PR_ROOT_DIR/.github/PULL_REQUEST_TEMPLATE.md")
 
   assert_contains "https://your-ticket-system.com/TICKET-123" "$actual"
 }
 
-function test_format_pr_body_link_with_branch_with_numbers_with_prefix() {
+function test_pr_body_link_with_branch_with_numbers_with_prefix() {
   export PR_TICKET_LINK_PREFIX=https://your-ticket-system.com/
 
-  local actual=$(format_pr_body "feat/TICKET-123-my-4-th-branch"\
+  local actual=$(pr_body "feat/TICKET-123-my-4-th-branch"\
     "$CREATE_PR_ROOT_DIR/.github/PULL_REQUEST_TEMPLATE.md")
 
   assert_contains "https://your-ticket-system.com/TICKET-123" "$actual"
 }
 
-function test_format_pr_body_background_with_link() {
+function test_pr_body_background_with_link() {
   export PR_TICKET_LINK_PREFIX=https://your-ticket-system.com/
 
-  local actual=$(format_pr_body "feat/TICKET-123-my-4-th-branch"\
+  local actual=$(pr_body "feat/TICKET-123-my-4-th-branch"\
     "$CREATE_PR_ROOT_DIR/.github/PULL_REQUEST_TEMPLATE.md")
 
   assert_not_contains "{{ BACKGROUND }}" "$actual"
@@ -108,10 +108,10 @@ function test_format_pr_body_background_with_link() {
   assert_contains "Details in the ticket." "$actual"
 }
 
-function test_format_pr_body_background_without_link() {
+function test_pr_body_background_without_link() {
   export PR_TICKET_LINK_PREFIX=
 
-  local actual=$(format_pr_body "feat/TICKET-123-my-4-th-branch"\
+  local actual=$(pr_body "feat/TICKET-123-my-4-th-branch"\
     "$CREATE_PR_ROOT_DIR/.github/PULL_REQUEST_TEMPLATE.md")
 
   assert_not_contains "{{ BACKGROUND }}" "$actual"