+++ tests/unit/pr_title_test.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 
 function set_up() {
-  source "$CREATE_PR_ROOT_DIR/src/pr_format.sh"
+  source "$CREATE_PR_ROOT_DIR/src/pr_title.sh"
 }
 
 function test_normalize_pr_title() {
@@ -28,34 +28,34 @@ function test_normalize_pr_title_with_prefix_and_ticket_number() {
   assert_same "Add pr 3 create Script" "$actual"
 }
 
-function test_format_title_without_ticket_number() {
-  actual="$(format_title "feat/TICKET-my-branch_name")"
+function test_pr_title_without_ticket_number() {
+  actual="$(pr_title "feat/TICKET-my-branch_name")"
 
   assert_same "Ticket my branch Name" "$actual"
 }
 
-function test_format_title_without_prefix() {
-  actual=$(format_title "TICKET-0000-add_pr_create_script")
+function test_pr_title_without_prefix() {
+  actual=$(pr_title "TICKET-0000-add_pr_create_script")
 
   assert_same "TICKET-0000 Add pr create script" "$actual"
 }
 
-function test_format_title_without_ticket_key() {
-  actual=$(format_title "0000-add_pr_create_script")
+function test_pr_title_without_ticket_key() {
+  actual=$(pr_title "0000-add_pr_create_script")
 
   assert_same "Add Pr Create Script" "$actual"
 }
 
-function test_format_title_without_ticket() {
-  actual=$(format_title "add-pr-create_script")
+function test_pr_title_without_ticket() {
+  actual=$(pr_title "add-pr-create_script")
 
   assert_same "Add pr create Script" "$actual"
 }
 
 # data_provider provider_no_prefix
-function test_format_title_remove_prefix() {
+function test_pr_title_remove_prefix() {
   local prefix=$1
-  actual=$(format_title "$prefix/TICKET-0000-my-new-3-feature")
+  actual=$(pr_title "$prefix/TICKET-0000-my-new-3-feature")
 
   assert_same "TICKET-0000 My new 3 feature" "$actual"
 }
@@ -67,25 +67,25 @@ function provider_no_prefix() {
 }
 
 # data_provider provider_fix_prefix
-function test_format_title_with_fix_prefix() {
+function test_pr_title_with_fix_prefix() {
   local prefix=$1
-  actual=$(format_title "$prefix/TICKET-0000-something-was-broken")
+  actual=$(pr_title "$prefix/TICKET-0000-something-was-broken")
 
   assert_same "TICKET-0000 Fix something was broken" "$actual"
 }
 
 # data_provider provider_fix_prefix
-function test_format_title_with_fix_prefix_and_bug_in_branch_name() {
+function test_pr_title_with_fix_prefix_and_bug_in_branch_name() {
   local prefix=$1
-  actual=$(format_title "$prefix/TICKET-0000-bug-something-was-broken")
+  actual=$(pr_title "$prefix/TICKET-0000-bug-something-was-broken")
 
   assert_same "TICKET-0000 Fix bug something was broken" "$actual"
 }
 
 # data_provider provider_fix_prefix
-function test_format_title_with_fix_prefix_and_fix_in_branch_name() {
+function test_pr_title_with_fix_prefix_and_fix_in_branch_name() {
   local prefix=$1
-  actual=$(format_title "$prefix/TICKET-0000-fix-something-was-broken")
+  actual=$(pr_title "$prefix/TICKET-0000-fix-something-was-broken")
 
   assert_same "TICKET-0000 Fix something was broken" "$actual"
 }